requirements.txt
fastmcp
PyJWT